inline mapQueryData transform
