+++ enterprise/frontend/src/embedding-sdk-package/hooks/public/use-metabase-query/map-dataset-query-data.ts
@@ -1,8 +1,8 @@
 import type { QueryDatasetResult } from "embedding-sdk-bundle/lib/query-dataset";
 import type { QueryQuestionResult } from "embedding-sdk-bundle/lib/query-question";
 
-import type { QueryData } from "../data-schema";
-import { mapRowsToObjects } from "../data-schema";
+import type { QueryData } from "../../../lib/public/data-schema";
+import { mapRowsToObjects } from "../../../lib/public/data-schema";
 
 export const mapDatasetQueryData = <TRow>(
   result: QueryQuestionResult | QueryDatasetResult,
