+++ enterprise/frontend/src/embedding-sdk-package/hooks/public/use-metabase-query/types.ts
@@ -13,7 +13,7 @@ import type {
   SchemaJavaScriptType,
   SchemaValue,
   TableSchema,
-} from "../data-schema";
+} from "../../../lib/public/data-schema";
 
 export type ID = string | number;
 type Values<T> = T[keyof T];
