+++ enterprise/frontend/src/embedding-sdk-package/lib/public/data-schema.ts
@@ -1,79 +1,34 @@
 import type { QueryQuestionResult } from "embedding-sdk-bundle/lib/query-question";
+import type {
+  FieldSchema,
+  MeasureSchema,
+  MetricSchema,
+  SchemaColumn,
+  SchemaJavaScriptType,
+  SchemaParameter,
+  SegmentSchema,
+  QuestionSchema as SharedQuestionSchema,
+  TableSchema,
+} from "embedding-sdk-shared/lib/create-metabase-query/schema";
 import type { DatasetColumn, RowValues } from "metabase-types/api";
+
+export type {
+  FieldSchema,
+  MeasureSchema,
+  MetricSchema,
+  SchemaColumn,
+  SchemaJavaScriptType,
+  SchemaParameter,
+  SegmentSchema,
+  TableSchema,
+};
 
 /**
  * A single value returned by Metabase query results or action responses.
  */
 export type RowValue = string | number | null | boolean | object;
 
-export type SchemaJavaScriptType =
-  | "string"
-  | "number"
-  | "boolean"
-  | "Date"
-  | "unknown";
-
-export type SchemaColumn = {
-  name: string;
-  description?: string;
-  displayName?: string;
-  jsType?: SchemaJavaScriptType;
-  baseType?: string;
-  effectiveType?: string;
-};
-
-export type SchemaParameter = {
-  slug: string;
-  type?: string;
-};
-
-export type QuestionSchema = {
-  id: QueryQuestionResult["id"];
-  columns: readonly SchemaColumn[];
-  parameters?: readonly SchemaParameter[];
-};
-
-/**
- * Metadata for a generated table field or metric dimension.
- */
-export type FieldSchema = SchemaColumn & {
-  id?: string | number;
-  fieldId?: number;
-  tableId?: number;
-  sourceFieldId?: number;
-};
-
-export type SegmentSchema<TTableId extends number = number> = {
-  kind: "segment";
-  id: number;
-  tableId: TTableId;
-};
-
-export type MeasureSchema<TTableId extends number = number> = {
-  kind: "measure";
-  id: number;
-  tableId: TTableId;
-  columns: readonly SchemaColumn[];
-};
-
-export type TableSchema = {
-  id: number;
-  databaseId: number;
-  columns?: readonly SchemaColumn[];
-  fields?: Record<string, FieldSchema>;
-  segments?: Record<string, SegmentSchema>;
-  measures?: Record<string, MeasureSchema>;
-};
-
-export type MetricSchema = {
-  id: number;
-  databaseId?: number;
-  sourceTableId?: number;
-  sourceCardId?: number;
-  columns: readonly SchemaColumn[];
-  mappedTableIds?: readonly number[];
-  dimensions?: Record<string, Record<string, FieldSchema>>;
-};
+export type QuestionSchema = SharedQuestionSchema<QueryQuestionResult["id"]>;
 
 export type SchemaValue<TColumn extends SchemaColumn> =
   TColumn["jsType"] extends "string"
@@ -138,16 +93,4 @@ export function mapRowsToObjects<TRow>(
       return acc;
     }, {}) as TRow;
   });
-}
-
-export function mapQueryData<TRow>(
-  result: QueryQuestionResult,
-): QueryData<TRow> {
-  const rawRows = result.rows;
-
-  return {
-    ...result,
-    rows: mapRowsToObjects<TRow>(result.columns, rawRows),
-    rawRows,
-  };
 }
